+++ torch/onnx/_internal/exporter/_torchlib/ops/nn.py
@@ -9,7 +9,6 @@ from __future__ import annotations
 from typing import Sequence, TYPE_CHECKING  # noqa: UP035
 
 from onnxscript.onnx_opset import (  # type: ignore[attr-defined]
-    opset18 as op18,
     opset20 as op20,
     opset21 as op21,
     opset23 as op23,
@@ -25,98 +24,6 @@ if TYPE_CHECKING:
     from onnxscript.values import Opset
 
 aten = torch.ops.aten
-
-_EINSUM_SYMBOLS = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
-
-
-def _get_einsum_symbol(dim: int) -> str:
-    if dim >= len(_EINSUM_SYMBOLS):
-        raise AssertionError(
-            "ONNX export for aten._trilinear only supports up to 52 dimensions"
-        )
-    return _EINSUM_SYMBOLS[dim]
-
-
-def _build_trilinear_subscript(total_dim: int, expanded_dims: Sequence[int]) -> str:
-    expanded_dims_set = set(expanded_dims)
-    return "".join(
-        _get_einsum_symbol(dim)
-        for dim in range(total_dim)
-        if dim not in expanded_dims_set
-    )
-
-
-def _build_trilinear_equation(
-    total_dim: int,
-    expand1: Sequence[int],
-    expand2: Sequence[int],
-    expand3: Sequence[int],
-    sumdim: Sequence[int],
-) -> str:
-    sumdim_set = set(sumdim)
-    output_subscript = "".join(
-        _get_einsum_symbol(dim)
-        for dim in range(total_dim)
-        if dim not in sumdim_set
-    )
-    return (
-        f"{_build_trilinear_subscript(total_dim, expand1)},"
-        f"{_build_trilinear_subscript(total_dim, expand2)},"
-        f"{_build_trilinear_subscript(total_dim, expand3)}->{output_subscript}"
-    )
-
-
-@onnx_impl(aten._trilinear.default, trace_only=True)
-def aten__trilinear(
-    i1: TReal,
-    i2: TReal,
-    i3: TReal,
-    expand1: Sequence[int],
-    expand2: Sequence[int],
-    expand3: Sequence[int],
-    sumdim: Sequence[int],
-    unroll_dim: int = 1,
-) -> TReal:
-    """_trilinear(Tensor i1, Tensor i2, Tensor i3, int[] expand1, int[] expand2, int[] expand3, int[] sumdim, int unroll_dim=1) -> Tensor"""
-
-    del unroll_dim
-
-    total_dim = len(i1.shape) + len(expand1)
-    equation = _build_trilinear_equation(total_dim, expand1, expand2, expand3, sumdim)
-    return op18.Einsum(i1, i2, i3, equation=equation)
-
-
-@onnx_impl(aten.bilinear.default, trace_only=True)
-def aten_bilinear(
-    input1: TReal,
-    input2: TReal,
-    weight: TReal,
-    bias: TReal | None = None,
-) -> TReal:
-    """bilinear(Tensor input1, Tensor input2, Tensor weight, Tensor? bias=None) -> Tensor"""
-
-    batch_shape = op18.Shape(input1, start=0, end=-1)
-    input1_shape = op18.Shape(input1, start=-1)
-    input2_shape = op18.Shape(input2, start=-1)
-    output_shape = op18.Shape(weight, start=0, end=1)
-    neg_1 = op18.Constant(value_ints=[-1])
-
-    weight_permuted = op18.Transpose(weight, perm=[1, 0, 2])
-    weight_flat = op18.Reshape(
-        weight_permuted,
-        op18.Concat(input1_shape, op18.Mul(output_shape, input2_shape), axis=0),
-    )
-
-    result = op18.MatMul(input1, weight_flat)
-    result = op18.Reshape(
-        result,
-        op18.Concat(batch_shape, output_shape, input2_shape, axis=0),
-    )
-    result = op18.Squeeze(op18.MatMul(result, op18.Unsqueeze(input2, neg_1)), neg_1)
-
-    if bias is not None:
-        result = op18.Add(result, bias)
-    return result
 
 
 @onnx_impl(aten.gelu.default, trace_only=True, opset_introduced=20)
